AGR-572  Common mapping across searchable documents
Updated mapping so that main fields for search/autocomplete will be generated by the same code across all of the categories, and threw in anything else that seemed to be common across docs.

I also got a start on a test that will run against ES to confirm that all of the categories are present, and that each field that should be loaded for documents in a category at least has one document with that field filled.